Basic flourish object tests
Flourish first gathers all TOML files found under the source directory
and creates an object for each. The test directory provides other types
of file to ensure they are not matched.

Test that we find what we are looking for, can get a single source
object by slug (which is its filename minus the `.toml` extension), and
that an exception is raised if the slug doesn’t match.