Flourish source lists can be sorted
The default sorting of `flourish.soures.all()` is filesystem dependent,
which is rarely helpful.

`order_by` takes a list of attributes to sort the objects on. If that
attribute is preceded by a hyphen (eg ‘-published’) then the sort is
reversed. If not all pages have the attribute, the sort will fail
with a warning.