Flourish objects can be filtered
`flourish.pages.filter()` implements basic filtering based upon keys
in the page’s configuration block. Numerical <, <=, >, and >= are
available, “contains” to find an item in an array or string, “in” to
find exact matches within a range of options, and “set” and “unset” to
see if a key exists (no matter the value).