Support Markdown content
Any source TOML file can have one or more Markdown blocks added to it
if they share the same slug part of the filename. For example, a
`blog-post.toml` file will read in `blog-post.summary.markdown` as
though it were specified as the key `summary_markdown` in the TOML.
Warnings are issued if this overwrites an existing key.

Any key in the resulting source configuration that ends `_markdown`
will be converted to HTML, eg `body_markdown` is converted and the 
result stored in `body`. Warnings are issued if this overwrites 
an existing key.