Support Markdown-only sources
A filename `slug.markdown` will be read in (whereas `slug.key.markdown`
will only try to attach to an existing source file). Markdown source
files can include TOML front matter by starting the file with three
hyphens (`---`) on a line, including the TOML, then finishing it with
three hyphens on a line. All other content is expected to be Markdown
and is stored in the key `body_markdown`.