Add URLs to a Flourish instance
URLs represent pages that Flourish will be able to generate from the current
source. New URLs need two parameters: the path, and a symbolic name with which
to reference the URL.

A hash (#) in the path of a URL represents a replaceable token, that will be
substituted as the output is generated. The alphanumerics that follow the hash
represent the key to use (as found in source TOML).