Add Atom feeds
Adds a new default Flourish generator `AtomGenerator` that will produce
an Atom feed from the current sources.

Introduces a new file to the source directory, `_site.toml`. This needs
three keys `author`, `title`, and `base_url` set in order for a valid
Atom feed to be generated. These, and any other values set in it are
also exposed to your templates in the default generators.

If there is no Atom feed to be generated, the `_site.toml` file is not
required.