Keep assets in source directory, not separately
As anything in the source directory that is not seen as a TOML, JSON
or Markdown source file was ignored. It makes more sense to treat them
as assets. This removes the need for a separate directory of “more stuff
for the website” and where a page is created from something such as an
image or video, there is now a more visible tie between the source file
and the associated media file(s).

As part of this, the logic for what URLs a Flourish instance has to
generate has moved into the Flourish instance from the command-line
runner code. This makes using Flourish via the API rather than the
command-line easier, and site generation is now more properly tested.