Make `lookup` template function more robust
Issue warnings when `lookup` fails, rather than raising. This means a
typo in a template won’t kill the preview server or a generate run.

Also renames `BaseSourceFile` to `SourceFile` and uses that instead of
`TomlSourceFile` as the exception class, because a missing source file
is not any specific flavour of file.