Rescan sources based on timestamps
When previewing a site with a large amount of source material,
creating a new Flourish object every time means tedious pauses
as each file is read, re-read, re-re-read...

Instead, read them once, and then rescan and re-read only those
files that have changed, reducing the time to generate the
resulting page significantly. Detecting changes is based on
file's timestamp, so it is not infallible.