Allow Atom feeds to be customised
Rather than insisting on the title and body to be exactly the same
as in the source, add methods to get that so they can be overridden.

The ID is most commonly the URL, but that does not have to be the
case, so allow that to be overridden too.