Test warnings in the objects tests
That Flourish generates warnings does not need to be tested repeatedly
in every set of tests. It (now) makes more sense to me that they are
tested the first time the sources are loaded in the test suite.

This makes using `pytest -x` easier, as source problems are found
before other problems that bad source loading will have a knock-on
effect upon.